requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.2
debugpy==1.6.5
decorator==5.1.1
dynaconf==3.1.11
entrypoints==0.4
executing==1.2.0
greenlet==2.0.2
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.9
jupyter_core==5.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pydantic==1.10.5
Pygments==2.14.0
python-dateutil==2.8.2
python-decouple==3.7
pywin32==305
pyzmq==25.0.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
stack-data==0.6.2
tinydb==4.7.0
tornado==6.2
traitlets==5.8.1
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
